Additive scene loading fix, future proofing(hopefully), Secret levels excluded from InLevel()
Agent found there to be an issue when using UltraTelephone with UltraKit since UltraKit loads scenes additively, potentially causing issues for plugins that use the level load/change events.
Since the code for the level load events in UltraTelephone is the exact same, it needs to be fixed here too.
Copied the fix Agent implemented in UltraTelephone to UKAPI.cs

* Added check for secret levels
* Added check for sanctum levels
* Prevents InLevel() returning true for secret levels
* Added CurrentSceneName static variable
* Added placeholder for upcoming credits level in UKLevelType